Add TryBothStores config option
Call Production App Store First. Then call sandbox. We're currently
doing this outside of the library which isn't ideal.
https://developer.apple.com/library/content/technotes/tn2259/_index.html

How do I verify my receipt (iOS)?

Always verify your receipt first with the production URL; proceed to
verify with the sandbox URL if you receive a 21007 status code.
Following this approach ensures that you do not have to switch between
URLs while your application is being tested or reviewed in the sandbox
or is live in the App Store.

PLAT-6523