Return errors from JSON encode/decode
I found two JSON errors being shadowed; the one in amazon/validator.go
is being hidden by errors.New(responseError.Message) -- which should be
an empty string if there's a JSON error. So, it wouldn't report success,
but this gives the caller better information on what failed.

The second is in appstore/validator.go, which was ignoring encode errors
before POSTing a verify request.